mesh_server/requirements.txt
flask>=2.0.0
requests>=2.25.0
PyJWT>=2.0.0
psycopg2-binary>=2.9.0
uuid>=1.30
lancedb>=0.3.0
pyarrow>=12.0.0